renamed files and fixed issues
